Move Tracker to unused Analytics class
